requirements.txt
tensorflow
matplotlib
Flask
Flask-Cors
Pillow
numpy
tf-keras-vis